Separate `forms.py` classes into different files
