+++ pyfirehose/config/ui/app.py
@@ -11,7 +11,8 @@ from npyscreen import NPSAppManaged
 from npyscreen import setTheme
 from npyscreen.npysThemes import DefaultTheme
 
-from pyfirehose.config.ui.forms.forms import MainForm, StubConfigEndpointsForm
+from pyfirehose.config.ui.forms.main import MainForm
+from pyfirehose.config.ui.forms.stub_config_edit import StubConfigEndpointsForm
 
 class ConfigApp(NPSAppManaged):
     """
