+++ pyfirehose/config/ui/forms/generic.py
@@ -1,0 +1,62 @@
+"""
+SPDX-License-Identifier: MIT
+"""
+
+from npyscreen import ActionFormV2, MiniButtonPress
+
+class ActionFormDiscard(ActionFormV2):
+    """
+    Generic class for an action form with an additional *Discard* button.
+
+    Overload the `on_discard` method to customize its behavior.
+    """
+    class Discard_Button(MiniButtonPress): #pylint: disable=invalid-name
+        """
+        Additional *Discard* button (name style chosen to match the `npyscreen` library).
+        """
+        def whenPressed(self):
+            self.parent._on_discard() #pylint: disable=protected-access
+
+    DISCARDBUTTON_TYPE = Discard_Button
+    DISCARD_BUTTON_BR_OFFSET = (
+        ActionFormV2.CANCEL_BUTTON_BR_OFFSET[0],
+        ActionFormV2.OK_BUTTON_BR_OFFSET[1]
+        + len(ActionFormV2.OK_BUTTON_TEXT)
+        + ActionFormV2.CANCEL_BUTTON_BR_OFFSET[1]
+        + len(ActionFormV2.CANCEL_BUTTON_TEXT)
+    )
+    DISCARD_BUTTON_TEXT = 'Discard'
+
+    def _on_discard(self):
+        self.editing = self.on_discard()
+
+    def create_control_buttons(self):
+        self._add_button('ok_button',
+            self.__class__.OKBUTTON_TYPE,
+            self.__class__.OK_BUTTON_TEXT,
+            0 - self.__class__.OK_BUTTON_BR_OFFSET[0],
+            0 - self.__class__.OK_BUTTON_BR_OFFSET[1] - len(self.__class__.OK_BUTTON_TEXT),
+            None
+        )
+
+        self._add_button('cancel_button',
+            self.__class__.CANCELBUTTON_TYPE,
+            self.__class__.CANCEL_BUTTON_TEXT,
+            0 - self.__class__.CANCEL_BUTTON_BR_OFFSET[0],
+            0 - self.__class__.CANCEL_BUTTON_BR_OFFSET[1] - len(self.__class__.CANCEL_BUTTON_TEXT),
+            None
+        )
+
+        self._add_button('discard_button',
+            self.__class__.DISCARDBUTTON_TYPE,
+            self.__class__.DISCARD_BUTTON_TEXT,
+            0 - self.__class__.DISCARD_BUTTON_BR_OFFSET[0],
+            0 - self.__class__.DISCARD_BUTTON_BR_OFFSET[1] - len(self.__class__.DISCARD_BUTTON_TEXT),
+            None
+        )
+
+    def on_discard(self):
+        """
+        *Discard* button hook to overload for customizing the behavior of the button.
+        """
+        return False
