+++ pyfirehose/config/ui/widgets/custom.py
@@ -2,15 +2,31 @@
 SPDX-License-Identifier: MIT
 
 Widgets used by forms to display and edit configuration files.
+
+--- Cansi library (https://github.com/tslight/cansi) ---
+Copyright (c) 2018, Toby Slight
+All rights reserved.
+
+Permission to use, copy, modify, and/or distribute this software for any purpose
+with or without fee is hereby granted, provided that the above copyright notice
+and this permission notice appear in all copies.
+
+THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
+REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND
+FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
+INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS
+OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER
+TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF
+THIS SOFTWARE.
+--- Cansi library (https://github.com/tslight/cansi) ---
 """
 
-import logging
 from typing import Optional
 
 import curses
+from npyscreen import Pager, SelectOne, Textfield, TitlePager, TitleSelectOne
 from pygments import highlight
 from pygments.formatters import TerminalFormatter #pylint: disable=no-name-in-module
-from npyscreen import Pager, SelectOne, Textfield, TitlePager, TitleSelectOne
 
 class CodeHighlightedTextfield(Textfield):
     """
